evaluateJavaScript to accept a shared_ptr instead of unique_ptr
Summary:
This changes jsi::evaluateJavaScript() to accept a
const shared_ptr<Buffer> & instead of a unique_ptr<Buffer.

It is reasonable to want to pass the same buffer to evaluateJavaScript()
multiple times. This will also help unify the API with the upcoming
prepareJavaScript() API.

Because shared_ptr has a unique_ptr constructor, this is compatible with
all call sites.

Reviewed By: mhorowitz

Differential Revision: D14001664

fbshipit-source-id: b7a0b7ec578a3fd6a6272241d50067269d2b03e4