Move unistd.h include to cpp where its used (#25107)
Summary:
unistd.h isn't a header available in the windows SDK, so we can't include it from react-native-windows.

I moved the usage of dup, to JSBigString.cpp in a previous PR, so this header should only be needed in the cpp file, not the header.  (And react-native-windows doesn't use the cpp file)

## Changelog

[Internal] [Fixed] - Header cleanup
Pull Request resolved: https://github.com/facebook/react-native/pull/25107

Differential Revision: D15602265

Pulled By: cpojer

fbshipit-source-id: 6a62bf8fe6758e400810f37834e8646485120d71