Do not load JS bundle from local cache if Metro is running
Summary: Having a cached JS bundle being loaded instead of loading it from Metro can cause recent changes made with Fast Refresh not to be picked up after an app is killed and started again. This diff makes a change so that we always download the bundle from Metro if it's running and only if it's not running we check if there is a cached recent bundle.

Reviewed By: gaearon, davidaurelio

Differential Revision: D16380243

fbshipit-source-id: b2842f718e0c21a3f4ca5ebeb3c2bae1df87a6e1