Adding the debug.keystore file back (#25807)
Summary:
Fixes https://github.com/facebook/react-native/issues/25629 that happened after v60 upgrade when using the diff tool https://react-native-community.github.io/upgrade-helper/?from=0.59.10&to=0.60.4

When following the diff for upgrading to react native 60. The debug.keystore is missing. It's added in the repository but ignored in the .gitignore, so it does not show. This adds an exception for this file.

## Changelog

[Android] [Fixed] - Add exception in .gitignore for `debug.keystore` to the android template.
Pull Request resolved: https://github.com/facebook/react-native/pull/25807

Test Plan: Create a new project from template and check that debug.keystore will be checked in when committing.

Reviewed By: cpojer

Differential Revision: D16462897

Pulled By: mdvacca

fbshipit-source-id: 18c0e8aee7f5cf8ed0d3544f88dc1621ad17f30f